Revamp slideshow section with dot styles, arrow shape options, and updated presets
