+++ app/sections/slideshow/dots.tsx
@@ -3,6 +3,7 @@ import { cva } from "class-variance-authority";
 import clsx from "clsx";
 
 export interface SlideshowDotsProps extends VariantProps<typeof variants> {
+  dotsStyle?: "circle" | "line" | "dash";
   className?: string;
 }
 
@@ -20,20 +21,44 @@ const variants = cva(
         right: "top-0! right-5! bottom-0! left-auto! flex-col",
       },
       dotsColor: {
-        light: "[&_.active]:outline-white! [&_.dot]:bg-white",
-        dark: "[&_.active]:outline-black! [&_.dot]:bg-black",
+        light: "[&_.dot]:bg-white",
+        dark: "[&_.dot]:bg-black",
+      },
+      dotsStyle: {
+        circle:
+          "[&_.active]:outline-current [&_.active]:outline-2 [&_.active]:outline-solid [&_.active]:outline-offset-3",
+        line: "[&_.active]:w-16 [&_.dot]:opacity-35 [&_.active]:opacity-100!",
+        dash: "[&_.dot]:opacity-35 [&_.active]:opacity-100!",
       },
     },
+    compoundVariants: [
+      {
+        dotsColor: "light",
+        dotsStyle: "circle",
+        className: "[&_.active]:outline-white!",
+      },
+      {
+        dotsColor: "dark",
+        dotsStyle: "circle",
+        className: "[&_.active]:outline-black!",
+      },
+    ],
     defaultVariants: {
       dotsPosition: "bottom",
       dotsColor: "light",
+      dotsStyle: "circle",
     },
   },
 );
 
 export function Dots(props: SlideshowDotsProps) {
-  const { className, dotsPosition, dotsColor } = props;
+  const { className, dotsPosition, dotsColor, dotsStyle } = props;
   return (
-    <div className={clsx(variants({ dotsPosition, dotsColor }), className)} />
+    <div
+      className={clsx(
+        variants({ dotsPosition, dotsColor, dotsStyle }),
+        className,
+      )}
+    />
   );
 }
